+++ tests/DefinitionTests.cs
@@ -67,8 +67,8 @@ namespace Worlds.Tests
 
             Assert.That(a.ComponentTypes.Count, Is.EqualTo(2));
             Assert.That(a.ArrayElementTypes.Count, Is.EqualTo(2));
-            USpan<ComponentType> componentTypes = stackalloc ComponentType[BitMask.Capacity];
-            USpan<ArrayElementType> arrayElementTypes = stackalloc ArrayElementType[BitMask.Capacity];
+            USpan<ComponentType> componentTypes = stackalloc ComponentType[(int)BitMask.Capacity];
+            USpan<ArrayElementType> arrayElementTypes = stackalloc ArrayElementType[(int)BitMask.Capacity];
             a.CopyComponentTypesTo(componentTypes);
             a.CopyArrayTypesTo(arrayElementTypes);
             Assert.That(componentTypes.Contains(schema.GetComponent<Integer>()), Is.True);
